Move checkout use of onStepChanged to Checkout component
